python-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.33.0
scipy==1.11.2
numpy==1.24.3
gunicorn==21.2.0
